add helper functions
- create_vpn_cert
Run easyrsa command that generates the certificate. This function users
pexpect which we may want to replace at some point.

- get_ovpn_file
run the openvpn command that exports the user details into and ovpn
file. This file can then be loaded directly into a VPN client.

- check_if_user_has_cert
check if a cert for the user already exits on disk

- check_password
does the password meet minimum requirements